refactor: Payment Builder 분리, 팩토리 메서드 유형별 분리, 상태전이 메서드 불변 네이밍 전환
