+++ bc/core/src/main/java/app/giftify/payment/application/PaymentFailService.java
@@ -20,7 +20,7 @@ public class PaymentFailService implements FailPaymentUseCase {
 	@Override
 	@Transactional
 	public void fail(Payment payment) {
-		Payment failed = payment.markAsFailed(LocalDateTime.now());
+		Payment failed = payment.fail(LocalDateTime.now());
 		var domainEvents = failed.pullEvents();
 		paymentRepository.save(failed);
 		domainEvents.forEach(eventPublisher::publish);
