+++ bc/core/src/test/java/app/giftify/payment/domain/PaymentMarkAsPartiallyCanceledTest.java
@@ -15,7 +15,7 @@ import app.giftify.shared.domain.type.PaymentMethod;
 import app.giftify.shared.domain.type.PaymentType;
 import app.giftify.shared.domain.vo.Money;
 
-@DisplayName("Payment.markAsPartiallyCanceled() 테스트")
+@DisplayName("Payment.partialCancel() 테스트")
 class PaymentMarkAsPartiallyCanceledTest {
 
 	private Payment createPaidPayment(Long paymentId, Money paidAmount) {
@@ -90,7 +90,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 			String reason = "부분 환불 요청";
 
 			// when
-			Payment result = payment.markAsPartiallyCanceled(newTransactionKey, cancelAmount, CancelType.REFUND, reason);
+			Payment result = payment.partialCancel(newTransactionKey, cancelAmount, CancelType.REFUND, reason);
 
 			// then
 			assertThat(result.getStatus()).isEqualTo(PaymentStatus.PARTIALLY_CANCELED);
@@ -109,7 +109,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 		}
 
 		@Test
-		@DisplayName("PAID 상태에서 전액 취소(markAsPartiallyCanceled 사용) → CANCELED 상태로 변경")
+		@DisplayName("PAID 상태에서 전액 취소(partialCancel 사용) → CANCELED 상태로 변경")
 		void fullCancelViaPartialCancel_FromPaid_ToCanceled() {
 			// given
 			Payment payment = createPaidPayment(1L, Money.of(10000));
@@ -118,7 +118,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 			String reason = "전액 환불";
 
 			// when
-			Payment result = payment.markAsPartiallyCanceled(newTransactionKey, cancelAmount, CancelType.REFUND, reason);
+			Payment result = payment.partialCancel(newTransactionKey, cancelAmount, CancelType.REFUND, reason);
 
 			// then
 			assertThat(result.getStatus()).isEqualTo(PaymentStatus.CANCELED);
@@ -136,7 +136,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 			String reason = "추가 부분 환불";
 
 			// when
-			Payment result = payment.markAsPartiallyCanceled(newTransactionKey, additionalCancelAmount, CancelType.REFUND, reason);
+			Payment result = payment.partialCancel(newTransactionKey, additionalCancelAmount, CancelType.REFUND, reason);
 
 			// then
 			assertThat(result.getStatus()).isEqualTo(PaymentStatus.PARTIALLY_CANCELED);
@@ -159,7 +159,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 			String reason = "최종 취소";
 
 			// when
-			Payment result = payment.markAsPartiallyCanceled(newTransactionKey, finalCancelAmount, CancelType.REFUND, reason);
+			Payment result = payment.partialCancel(newTransactionKey, finalCancelAmount, CancelType.REFUND, reason);
 
 			// then
 			assertThat(result.getStatus()).isEqualTo(PaymentStatus.CANCELED);
@@ -181,7 +181,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 
 			// when & then
 			assertThatThrownBy(() ->
-				payment.markAsPartiallyCanceled("txn-excessive", excessiveCancelAmount, CancelType.REFUND, "초과 환불")
+				payment.partialCancel("txn-excessive", excessiveCancelAmount, CancelType.REFUND, "초과 환불")
 			)
 				.isInstanceOf(PaymentException.class)
 				.extracting("errorCode")
@@ -197,7 +197,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 
 			// when & then
 			assertThatThrownBy(() ->
-				payment.markAsPartiallyCanceled("txn-excessive", excessiveCancelAmount, CancelType.REFUND, "초과 환불")
+				payment.partialCancel("txn-excessive", excessiveCancelAmount, CancelType.REFUND, "초과 환불")
 			)
 				.isInstanceOf(PaymentException.class)
 				.extracting("errorCode")
@@ -213,7 +213,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 
 			// when & then
 			assertThatThrownBy(() ->
-				payment.markAsPartiallyCanceled("txn-invalid", cancelAmount, CancelType.REFUND, "대기 중 취소")
+				payment.partialCancel("txn-invalid", cancelAmount, CancelType.REFUND, "대기 중 취소")
 			)
 				.isInstanceOf(PaymentException.class)
 				.extracting("errorCode")
@@ -235,7 +235,7 @@ class PaymentMarkAsPartiallyCanceledTest {
 			String reason = "환불 테스트";
 
 			// when
-			Payment result = payment.markAsPartiallyCanceled(transactionKey, cancelAmount, CancelType.REFUND, reason);
+			Payment result = payment.partialCancel(transactionKey, cancelAmount, CancelType.REFUND, reason);
 
 			// then
 			List<Object> events = result.pullEvents();
